requirements.txt
aiohttp>=3.01
aiohttp-basicauth-middleware==1.1.2
colorlog
dnspython==1.16.0
